Make the tests immune to the locale.  (LP: #1638570)
