+++ debian/changelog
@@ -7,6 +7,7 @@ ubuntu-image (0.13+17.04ubuntu1) UNRELEASED; urgency=medium
     this can fail.  Over local file systems, they'll be preserved anyway.
     (LP: #1637554)
   * d/tests/mount: Switch to the stable channel for snaps.
+  * Make the tests immune to the locale.  (LP: #1638570)
 
  -- Barry Warsaw <barry@ubuntu.com>  Thu, 01 Dec 2016 17:33:19 -0500
